Refactoring Code into a new DAO class and SQLConstants class
