+++ components/apimgt/org.wso2.carbon.apimgt.cleanup.service/src/main/java/org/wso2/carbon/apimgt/cleanup/service/IdpKeyMangerPurge.java
@@ -45,12 +45,12 @@ public class IdpKeyMangerPurge implements OrganizationPurge {
     protected String username;
     APIAdmin apiAdmin;
     LinkedHashMap<String, String> IdpKeyMangerPurgeTaskMap = new LinkedHashMap<>();
-    ApiMgtDAO apiMgtDAO;
+    OrganizationPurgeDAO organizationPurgeDAO;
     private static final Log log = LogFactory.getLog(IdpKeyMangerPurge.class);
 
     public IdpKeyMangerPurge(String username) {
         this.username = username;
-        apiMgtDAO = ApiMgtDAO.getInstance();
+        organizationPurgeDAO = OrganizationPurgeDAO.getInstance();
         this.apiAdmin = new APIAdminImpl();
         initTaskList();
     }
@@ -64,7 +64,7 @@ public class IdpKeyMangerPurge implements OrganizationPurge {
                 .put(APIConstants.OrganizationDeletion.KM_DATA_REMOVER, APIConstants.OrganizationDeletion.PENDING);
     }
 
-    public LinkedHashMap<String, String> deleteOrganization(String organization) {
+    public LinkedHashMap<String, String> purge(String organization) {
 
         List<KeyManagerConfigurationDTO> keyManagerList = new ArrayList<>();
         for (Map.Entry<String, String> task : IdpKeyMangerPurgeTaskMap.entrySet()) {
@@ -80,7 +80,7 @@ public class IdpKeyMangerPurge implements OrganizationPurge {
                         deleteIdpList(organization, keyManagerList);
                         break;
                     case APIConstants.OrganizationDeletion.KM_DATA_REMOVER:
-                        apiMgtDAO.deleteKeyManagerConfigurationList(keyManagerList, organization);
+                        organizationPurgeDAO.deleteKeyManagerConfigurationList(keyManagerList, organization);
                         break;
                     }
                     IdpKeyMangerPurgeTaskMap.put(task.getKey(), APIConstants.OrganizationDeletion.COMPLETED);
